[MIST-651] Move classes related to query merging into merging package
